@@ -13,49 +13,49 @@
     "build:css": "yarn postcss",
     "build": "NODE_ENV=production concurrently \"yarn build:js\" \"yarn build:css\"",
     "watch:js": "yarn esbuild --watch",
-    "watch:css": "TAILWIND_MODE=watch yarn postcss -w",
+    "watch:css": "yarn postcss -w",
     "watch": "concurrently --kill-others \"yarn watch:js\" \"yarn watch:css\"",
     "start": "concurrently --kill-others \"yarn esbuild --servedir=www\" \"yarn watch:css\"",
     "test": "jest",
     "test:coverage": "jest --collectCoverage"
   },
   "devDependencies": {
-    "@testing-library/dom": "^8.10.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.3",
-    "@types/react": "^17.0.31",
-    "@types/react-dom": "^17.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
-    "autoprefixer": "^10.3.7",
-    "concurrently": "^6.3.0",
-    "cssnano": "^5.0.8",
-    "esbuild": "^0.13.8",
-    "eslint": "^8.0.1",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.2",
+    "@testing-library/dom": "^8.13.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^16.11.32",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.21.0",
+    "@typescript-eslint/parser": "^5.21.0",
+    "autoprefixer": "^10.4.5",
+    "concurrently": "^7.1.0",
+    "cssnano": "^5.1.7",
+    "esbuild": "^0.14.38",
+    "eslint": "^8.14.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "jest": "^27.3.1",
-    "msw": "^0.35.0",
-    "postcss": "^8.3.11",
-    "postcss-cli": "^9.0.1",
-    "postcss-import": "^14.0.2",
-    "postcss-nesting": "^8.0.1",
-    "prettier": "^2.4.1",
-    "tailwindcss": "^2.2.17",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.4.4",
+    "eslint-plugin-react": "^7.29.4",
+    "jest": "^28.0.3",
+    "msw": "^0.39.2",
+    "postcss": "^8.4.13",
+    "postcss-cli": "^9.1.0",
+    "postcss-import": "^14.1.0",
+    "postcss-nesting": "^10.1.4",
+    "prettier": "^2.6.2",
+    "tailwindcss": "^3.0.24",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.4",
     "whatwg-fetch": "^3.6.2"
   },
   "dependencies": {
     "classnames": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   },
-  "packageManager": "yarn@3.0.2"
+  "packageManager": "yarn@3.2.0"
 }
