@@ -1,6 +1,5 @@
 module.exports = {
-  mode: 'jit',
-  purge: ['./src/**/*.ts', './src/**/*.tsx', './www/index.html'],
+  content: ['./src/**/*.ts', './src/**/*.tsx', './www/index.html'],
   darkMode: 'media',
   theme: {
     extend: {},
